+++ packages/presets/client/tests/client-preset.spec.ts
@@ -619,10 +619,7 @@ export * from "./gql";`);
       export type MeQueryVariables = Exact<{ [key: string]: never; }>;
 
 
-      export type MeQuery = { __typename?: 'Query', unmasked?: (
-          { __typename?: 'User', id: string, name: string, age: number }
-          & { ' $fragmentRefs'?: { 'User_MeFragment': User_MeFragment } }
-        ) | null, masked?: (
+      export type MeQuery = { __typename?: 'Query', unmasked?: { __typename?: 'User', id: string, name: string, age: number } | null, masked?: (
           { __typename?: 'User', id: string }
           & { ' $fragmentRefs'?: { 'User_MeFragment': User_MeFragment } }
         ) | null };
@@ -2467,9 +2464,13 @@ export * from "./gql.js";`);
           implements DocumentTypeDecoration<TResult, TVariables>
         {
           __apiType?: DocumentTypeDecoration<TResult, TVariables>['__apiType'];
+          private value: string;
+          public __meta__?: Record<string, any> | undefined;
 
-          constructor(private value: string, public __meta__?: Record<string, any> | undefined) {
+          constructor(value: string, __meta__?: Record<string, any> | undefined) {
             super(value);
+            this.value = value;
+            this.__meta__ = __meta__;
           }
 
           toString(): string & DocumentTypeDecoration<TResult, TVariables> {
@@ -2608,9 +2609,13 @@ export * from "./gql.js";`);
           implements DocumentTypeDecoration<TResult, TVariables>
         {
           __apiType?: DocumentTypeDecoration<TResult, TVariables>['__apiType'];
+          private value: string;
+          public __meta__?: Record<string, any> | undefined;
 
-          constructor(private value: string, public __meta__?: Record<string, any> | undefined) {
+          constructor(value: string, __meta__?: Record<string, any> | undefined) {
             super(value);
+            this.value = value;
+            this.__meta__ = __meta__;
           }
 
           toString(): string & DocumentTypeDecoration<TResult, TVariables> {
@@ -2744,9 +2749,13 @@ export * from "./gql.js";`);
           implements DocumentTypeDecoration<TResult, TVariables>
         {
           __apiType?: DocumentTypeDecoration<TResult, TVariables>['__apiType'];
+          private value: string;
+          public __meta__?: Record<string, any> | undefined;
 
-          constructor(private value: string, public __meta__?: Record<string, any> | undefined) {
+          constructor(value: string, __meta__?: Record<string, any> | undefined) {
             super(value);
+            this.value = value;
+            this.__meta__ = __meta__;
           }
 
           toString(): string & DocumentTypeDecoration<TResult, TVariables> {
@@ -3039,9 +3048,13 @@ export * from "./gql.js";`);
           implements DocumentTypeDecoration<TResult, TVariables>
         {
           __apiType?: DocumentTypeDecoration<TResult, TVariables>['__apiType'];
+          private value: string;
+          public __meta__?: Record<string, any> | undefined;
 
-          constructor(private value: string, public __meta__?: Record<string, any> | undefined) {
+          constructor(value: string, __meta__?: Record<string, any> | undefined) {
             super(value);
+            this.value = value;
+            this.__meta__ = __meta__;
           }
 
           toString(): string & DocumentTypeDecoration<TResult, TVariables> {